requirements.txt
fastapi[all]
prometheus-client
environ-config
pydantic
uvicorn
sqlalchemy
psycopg2-binary
httpx
